Changed lons to long longs. longs only 4 bytes in Visual Studio. 8 bytes in Linux Mint

+++ bi_type.cpp
@@ -35,9 +35,9 @@ static char vcid[] = "$Id: bi_type.c,v 1.2 1994/12/08 23:08:52 duchier Exp $";
 /******** C_CHILDREN
   Return a list of roots of the children types of T (except bottom).
 */
-static long c_children()   /*  RM: Dec 14 1992  Re-wrote most of the routine */
+static long long c_children()   /*  RM: Dec 14 1992  Re-wrote most of the routine */
 {
-  long success=TRUE;
+  long long success=TRUE;
   ptr_psi_term funct,result,arg1,arg2,t,p1;
   ptr_int_list p;
   
@@ -98,9 +98,9 @@ static long c_children()   /*  RM: Dec 14 1992  Re-wrote most of the routine */
 /******** C_PARENTS
   Return a list of roots of the parent types of T.
 */
-static long c_parents()
+static long long c_parents()
 {
-  long success=TRUE;
+  long long success=TRUE;
   ptr_psi_term funct,result,arg1,arg2,t,p1;
   ptr_int_list p;
 
@@ -123,7 +123,7 @@ static long c_parents()
     else {
       if ((arg1->type==quoted_string || arg1->type==integer ||
           arg1->type==real) && arg1->value_3!=NULL) {
-        /* arg1 is a string, long or real: return a list with arg1 as
+        /* arg1 is a string, long long or real: return a list with arg1 as
            argument, where arg1->value = NULL, MH */
         p1 = stack_psi_term(4);
         p1->type = arg1->type;
@@ -158,9 +158,9 @@ static long c_parents()
   Return the parents of bottom.
   This function has no arguments.
 */
-static long c_smallest()
+static long long c_smallest()
 {
-  long success=TRUE;
+  long long success=TRUE;
   ptr_psi_term result, g, t;
 
   g=aim->aaaa_1;
@@ -172,13 +172,13 @@ static long c_smallest()
   return success;
 }
 
-long isSubTypeValue(ptr_psi_term arg1, ptr_psi_term arg2)  // REV401PLUS changed to long
+long long isSubTypeValue(ptr_psi_term arg1, ptr_psi_term arg2)  // REV401PLUS changed to long long
 // ptr_psi_term arg1, arg2;
 {
-  long ans=TRUE;
+  long long ans=TRUE;
   
   /* we already know that either arg1->type == arg2->type or that at both
-   * of the two are either long or real
+   * of the two are either long long or real
    */
   
   if (arg2->value_3) {
@@ -205,10 +205,10 @@ long isSubTypeValue(ptr_psi_term arg1, ptr_psi_term arg2)  // REV401PLUS changed
 }
 
 /* Boolean utility function that implements isa */
-static long isa(ptr_psi_term arg1, ptr_psi_term arg2)
+static long long isa(ptr_psi_term arg1, ptr_psi_term arg2)
 // ptr_psi_term arg1, arg2;
 {
-  long ans;
+  long long ans;
 
   if (  arg1->type==arg2->type
      || (  (arg1->type==real || arg1->type==integer)
@@ -246,11 +246,11 @@ static long isa(ptr_psi_term arg1, ptr_psi_term arg2)
 #define isa_ncmp_sel 11
 
 /* Utility that selects one of several isa functions */
-static long isa_select(ptr_psi_term arg1, ptr_psi_term arg2, long sel)
+static long long isa_select(ptr_psi_term arg1, ptr_psi_term arg2, long long sel)
 // ptr_psi_term arg1,arg2;
-// long sel;
+// long long sel;
 {
-  long ans;
+  long long ans;
 
   switch (sel) {
   case isa_le_sel: ans=isa(arg1,arg2);
@@ -286,10 +286,10 @@ static long isa_select(ptr_psi_term arg1, ptr_psi_term arg2, long sel)
 /******** C_ISA_MAIN
   Main routine to handle all the isa built-in functions.
 */
-static long c_isa_main(long sel)
-// long sel;
+static long long c_isa_main(long long sel)
+// long long sel;
 {
-  long success=TRUE,ans;
+  long long success=TRUE,ans;
   ptr_psi_term arg1,arg2,funct,result;
 
   funct=aim->aaaa_1;
@@ -316,62 +316,62 @@ static long c_isa_main(long sel)
   and lists.  For lists, it looks only at the top level, i.e. whether the
   object is nil or a cons cell.
 */
-static long c_isa_le()
+static long long c_isa_le()
 {
   return c_isa_main(isa_le_sel);
 }
 
-static long c_isa_lt()
+static long long c_isa_lt()
 {
   return c_isa_main(isa_lt_sel);
 }
 
-static long c_isa_ge()
+static long long c_isa_ge()
 {
   return c_isa_main(isa_ge_sel);
 }
 
-static long c_isa_gt()
+static long long c_isa_gt()
 {
   return c_isa_main(isa_gt_sel);
 }
 
-static long c_isa_eq()
+static long long c_isa_eq()
 {
   return c_isa_main(isa_eq_sel);
 }
 
-static long c_isa_nle()
+static long long c_isa_nle()
 {
   return c_isa_main(isa_nle_sel);
 }
 
-static long c_isa_nlt()
+static long long c_isa_nlt()
 {
   return c_isa_main(isa_nlt_sel);
 }
 
-static long c_isa_nge()
+static long long c_isa_nge()
 {
   return c_isa_main(isa_nge_sel);
 }
 
-static long c_isa_ngt()
+static long long c_isa_ngt()
 {
   return c_isa_main(isa_ngt_sel);
 }
 
-static long c_isa_neq()
+static long long c_isa_neq()
 {
   return c_isa_main(isa_neq_sel);
 }
 
-static long c_isa_cmp()
+static long long c_isa_cmp()
 {
   return c_isa_main(isa_cmp_sel);
 }
 
-static long c_isa_ncmp()
+static long long c_isa_ncmp()
 {
   return c_isa_main(isa_ncmp_sel);
 }
@@ -381,7 +381,7 @@ static long c_isa_ncmp()
 /******** C_IS_FUNCTION
   Succeed iff argument is a function (built-in or user-defined).
 */
-static long c_is_function() /*  RM: Jan 29 1993  */ // REV401PLUS long
+static long long c_is_function() /*  RM: Jan 29 1993  */ // REV401PLUS long long
 {
   int success=TRUE,ans;
   ptr_psi_term arg1,funct,result;
@@ -407,7 +407,7 @@ static long c_is_function() /*  RM: Jan 29 1993  */ // REV401PLUS long
 /******** C_IS_PERSISTENT
   Succeed iff argument is a quoted persistent or on the heap.
 */
-static long c_is_persistent() /*  RM: Feb  9 1993  */ // REV401PLUS long
+static long long c_is_persistent() /*  RM: Feb  9 1993  */ // REV401PLUS long long
 {
   int success=TRUE,ans;
   ptr_psi_term arg1,glob,result;
@@ -437,7 +437,7 @@ static long c_is_persistent() /*  RM: Feb  9 1993  */ // REV401PLUS long
 /******** C_IS_PREDICATE
   Succeed iff argument is a predicate (built-in or user-defined).
 */
-static long c_is_predicate() /*  RM: Jan 29 1993  */ // REV401PLUS long
+static long long c_is_predicate() /*  RM: Jan 29 1993  */ // REV401PLUS long long
 {
   int success=TRUE,ans;
   ptr_psi_term arg1,funct,result;
@@ -463,7 +463,7 @@ static long c_is_predicate() /*  RM: Jan 29 1993  */ // REV401PLUS long
 /******** C_IS_SORT
   Succeed iff argument is a sort (built-in or user-defined).
 */
-static long c_is_sort() /*  RM: Jan 29 1993  */ // REV401PLUS long
+static long long c_is_sort() /*  RM: Jan 29 1993  */ // REV401PLUS long long
 {
   int success=TRUE,ans;
   ptr_psi_term arg1,funct,result;
@@ -491,9 +491,9 @@ static long c_is_sort() /*  RM: Jan 29 1993  */ // REV401PLUS long
   a quirky way in Wild_Life.  This is true for integers, reals,
   strings (which are potentially infinite sets of objects), and list objects.
 */
-static long c_is_value()
+static long long c_is_value()
 {
-  long success=TRUE,ans;
+  long long success=TRUE,ans;
   ptr_psi_term arg1,arg2,funct,result;
 
   funct=aim->aaaa_1;
@@ -516,9 +516,9 @@ static long c_is_value()
 /******** C_IS_NUMBER
   Return true iff argument is an actual number.
 */
-static long c_is_number()
+static long long c_is_number()
 {
-  long success=TRUE,ans;
+  long long success=TRUE,ans;
   ptr_psi_term arg1,arg2,funct,result;
 
   funct=aim->aaaa_1;
@@ -541,7 +541,7 @@ static long c_is_number()
   if A is a subsort of B => succeed and residuate on B
   else			 => fail
 */
-long c_isa_subsort() // changed to long REV401PLUS
+long long c_isa_subsort() // changed to long long REV401PLUS
 {
   ptr_psi_term pred,arg1,arg2;
 
@@ -567,7 +567,7 @@ long c_isa_subsort() // changed to long REV401PLUS
 
 
 
-long isValue(ptr_psi_term p)  // REV401PLUS to long
+long long isValue(ptr_psi_term p)  // REV401PLUS to long long
 // ptr_psi_term p;
 {
 	return (p->value_3 != NULL);
@@ -579,13 +579,13 @@ long isValue(ptr_psi_term p)  // REV401PLUS to long
   Return glb(A,B).  Continued calls will return each following type in
   the disjunction of the glb of A,B.
 */
-long c_glb()  // REV401PLUS long
+long long c_glb()  // REV401PLUS long long
 {
   ptr_psi_term func,arg1,arg2, result, other;
   ptr_definition ans;
   ptr_int_list complexType;
   ptr_int_list decodedType = NULL;
-  long ret;
+  long long ret;
   
   func=aim->aaaa_1;
   deref_ptr(func);
@@ -640,7 +640,7 @@ long c_glb()  // REV401PLUS long
   Return lub(A,B).  Continued calls will return each following type in
   the disjunction of the lub of A,B.
 */
-long c_lub()  // REV401PLUS long
+long long c_lub()  // REV401PLUS long long
 {
   ptr_psi_term func,arg1,arg2, result, other;
   ptr_definition ans=NULL;
